Replace AILevel with IVs in TrainerPokemon (#195)
* Replace AILevel with IVs

AILevel was a bit misleading, so this updates the values.

IVs were checked by setting up a fight against a level 100 pokemon with endeavor as the only move. This way, we can observe the HP of the trainer pokemon and derive the IVs given the known base stats.

Along the way, I figured out a few more of the trainer fields that may be useful in future features, so I documented what I found.

The only real snag is the difference in the "strength" or "IV" field in gen 6 between ORAS and X/Y. To account for the unknown purpose of the high 3 bits in X/Y, we add a field to the TrainerPokemon class to account for that.